Break up exceptions.
