+++ src/main/java/com/nepherte/commons/cli/exception/ExclusiveOptionsException.java
@@ -1,0 +1,37 @@
+/*
+ * Copyright 2012-2020 Bart Verhoeven
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.nepherte.commons.cli.exception;
+
+import com.nepherte.commons.cli.Option;
+
+import java.util.Arrays;
+
+/**
+ * Exception thrown when a command has mutually exclusive options.
+ */
+public final class ExclusiveOptionsException extends ParseException {
+
+  private static final long serialVersionUID = 356042360910950845L;
+
+  /**
+   * Creates a new {@code ExclusiveOptionsException}.
+   *
+   * @param templates the templates of the mutually exclusive options
+   */
+  public ExclusiveOptionsException(Option.Template... templates) {
+    super("Mutually exclusive options " + Arrays.toString(templates));
+  }
+}
